Add export-to-CSV functionality
There is now a link on the sidebar under "Tournament report", labelled
"CSV". It leads you to a page which asks you for the event code and
game format as it should appear in the CSV file. It will then generate
you the games in the same CSV format as our massive Google Docs
spreadsheet requires. This makes copying event results into the
spreadsheet far easier.

Note that Atropine doesn't know about the format of the game, whether
it's a 9 rounder, 15 rounder, or whatever. (It tries not to assume
you're even playing Countdown.) You can specify one "format" for the
whole CSV file, e.g. "9R" or "15R". If different games in the same
tournament should have different format codes, you'll need to fix that
manually.