Round automatic ratings to two decimal places
When we automatically assign ratings, round them to two decimal places.
This is because when we show the Edit Player pane in the Player page,
the rating shown there is rounded to two decimal places. So if Fred's
rating is 1166.666666, it's displayed in the edit box as 1166.67. If we
change something else and save the changes, the page will give a
confusing message suggesting it changed the rating as well - which it
did, from 1166.666666 to 1166.67.

So let's make all automatically-assigned ratings rounded to two decimal
places to avoid this.