Enhance CSV textarea in Tourney Setup
Currently, the textarea on the Tourney Setup page contains a list of
player names, each optionally followed by a rating.

Alice,2000
Bob,1900
Charlie,1800
... etc

When you submit the player list on this page, it deletes any existing
players and recreates them. This means any other settings you've made
for those players, such as whether they require an accessible table, or
what team they're on if any, are lost.

So let's enhance the format expected in this textarea in a
backwardly-compatible way, as follows.

The textarea contains CSV.

In each record, the first field is the player's name, as before. Also as
before, if this is a dash ("-") it indicates a division boundary, it
doesn't represent a player, and any subsequent fields are ignored.

There may be any number of CSV fields after the player's name. Their
order is not significant, but fields of specific strings have special
meanings as follows.

For each field after the player name:
* If a field parses as a number, it is the player's rating. As before,
  a rating of 0 indicates a Prune.
* If the field is "A" then the player requires an accessible table.
* If the field is "NP" then the player avoids Prune.
* If the field is "W" then the player is withdrawn (yes, withdrawn
  before the tourney even started). Possibly useful if you know a
  player is going to arrive late and intend to unwithdraw them later on.
* If the field is "T1" or "T2" then the player is on team 1 or 2.
* Otherwise, the field is ignored.

For example, you could have a player list like this, which indicates
Charlie and Edward require accessible tables, and Edward isn't to be
drawn against Prune:

Alice,2000
Bob,1900,
Charlie,1800,A
Dave,1700
Edward,1600,A,NP
Fred,1500
George,1400
Prune,0

If you change Team assignments, player attributes, etc, before
generating any fixtures, then go back to the Tourney Setup page, the CSV
in the textarea should reflect the changes you've made, so if you make a
minor change to this text and resubmit it, your previous changes will
still be there.