Don't show withdrawn players with no games in the standings table
If a player is withdrawn, and they have not played any games, then
there's no real point in displaying them in the public-facing standings
table. This is especially true if this player would take the number of
players in the table to 1 more than a multiple of 8, because it would
mean the public display's standings table has an entire useless page
showing one withdrawn player on zero games.

This situation might arise if someone starts the tournament as
withdrawn because they aren't there, and has their place taken by a
Prune, but never turns up.

Withdrawn players who have not played any games will now no longer be
included in the standings table on the public display, or on the
standings table shown in any of the "export" pages.