Changes to player list textarea interface on Tourney Setup page
If no games have been generated, you get a textarea on the Tourney Setup
page into which you can paste, or type, the player list.

There is now an automatically-updating label below the textarea, telling
you how many active players are listed, and, if there is more than one
division, how many players in each division. This is useful if you've
pasted in a player list from somewhere else and need to know before
saving it how many players that is and so whether you need to add a
Prune or two.

The help box next to the textarea now has an explanation of the "A",
"NP" and "W" flags you can put after each player's name. It is initially
hidden behind a "show more" link.