Tim Down Award page
As well as Tuff Luck and Overachievers pages, there is now a Tim Down
Award page. The winner of the Tim Down award is the person whose
opponents have the highest average position in the standings table,
and who lost three or more games.

There isn't a Teleost view for this, but there's no reason why there
couldn't be one.