Allow players to prefer specific tables
As well as requiring an accessible table, a player can now be assigned
a "preferred" table, and the player will be seated at that table if
possible, subject to other players' preferences and accessibility
requirements.

This is in case there are two players who require accessible tables,
but one of these players prefers to sit at table 1 while the other
prefers table 20. Previously you could set tables 1 and 20 as
accessible tables but there'd be no guarantee that those players
would be assigned to those tables the "right" way round.

The group-to-table assignment logic has been moved to cttable.py.
Each group has a preferred table and a preference strength, which
is the sum of its players' preferred tables and preference strengths.
A player requiring a specific accessible table has a greater weight
than a player requiring any accessible tbale, which in turn has a
greater weight than a player requiring a specific table but not
due to accessibility requirements.

Table are then given to groups in order of the groups' preference
strengths. The first group is asked which table it prefers, then the
next group is asked which table it prefers out of the remaining
tables, and so on.

If there are no preferences, or a group's preference is equally split
between multiple tables, then a group will prefer table numbers from
the natural range that would be assigned to that division, and if
there's still a decision to be made, it will pick the lowest-numbered
acceptable table.

If two groups' preference strengths are the same, we give first
pick to the group which appeared first in the list the fixture
generator gave us. This means if there are no preferences or
accessibility requirements, the table numbers will be assigned in
the order they came out of the fixture generator, and so the
convention that higher-ranked players go on lower-numbered tables
(for the Swiss fixture generator) is preserved.