Improvements to table preferences
If you set a player's table preference to something that isn't an
accessible table, and the player requires an accessible table, you
now get a warning.

cttable.py: take into account that after some tables have been
assigned, a TableVotingGroup that originally had a preference
might no longer have a preference (all its preferred table have
gone) so will need to be transferred to the "ordinary" groups
list, which are given table numbers in the order the groups
came out of the fixture generator.

Small formatting changes to the player setup page.