User interface changes to player setup and general setup
The list of players in the Player Setup page now shows at a glance
which players have a preferred table number, by showing the table
number alongside the player on a blue background.

The General Setup page has had some minor layout tweaks, including
labels for radio buttons now actually being in <label> tags so that
the user can now click the label in place of the radio button.