Output sidebar and mainpane after generating fixtures
Redo fixturegen.py so that we do whatever's requested, *then* output
the sidebar and main HTML, rather than writing out bits and bobs as
we go along. This is so that when the user generates fixtures for a
new round and presses "Accept Fixtures", they're taken to a page which
has a link for the new round in the sidebar. If we print the sidebar
before we write the fixtures to the db, the sidebar won't contain the
new round.

Also tweak the way fixture generators' radio buttons and check boxes
are shown, so that their label is actually an HTML <label> and can be
clicked to use the control.