Potential future changes to public display style
The teleoststyle.css now has the Roboto font defined in it, making it
easier to switch from Roboto Condensed to Roboto if desired.

The Standings view now uses an option standings_alternate_row_colours,
which, if set, causes the rows in the standings table to have
alternating row colours for ease of reading. Currently there's no way
to actually switch on this option besides manually running SQL in the
db, but it's there in case we ever want to add the option.