Tweaks and fixes to player configuration page
Minor tweaks to the player configuration page: make the edit boxes
for "rating" and "preferred table" shorter, and add a line of
explanation for "rating".

Also fix a bug where you could rename a player to a string which
matched the name of another player case-insensitively but not
case-sensitively, and you wouldn't get an exception. Instead you'd
get an ugly stack trace caused by violating a uniqueness constraint
in the database. countdowntourney.rename_player() now checks the
proposed new name against existing player names case-insensitively,
and throws the expected exception if there's a match.