Raw SQL interface, for debugging and emergency surgery
From the "Advanced" tourney setup page, you can now get to sql.py,
which allows you to run SQL on the tourney database. There are
suitably scary warnings against messing around with it.

The Raw SQL Interface page shows the available tables and views in
the sidebar. There is a textarea to enter SQL, and any results are
displayed in a table below.

sql.py deliberately does not use the countdowntourney object to
load the db file like everything else does, because if you're in the
situation where you're using sql.py, this might be because the
countdowntourney code doesn't like something about the database
(table missing, unexpected value in a column, etc) so might refuse
to use it. In theory this means you could use sql.py to run SQL
on any SQLite database - it doesn't have to be an atropine tourney
database.

Also made a few UI tweaks - e.g. on delround.py, made the cancel
button a neutral colour rather than a possibly misleading green.