Option to show unbelievably high scores in words on the videprinter
For no reason, sports videprinters like to follow particularly
unbelievable scores with the number spelt out in words, e.g.
Some Team 1 Another Team 7 (seven)

This feature has been requested for Atropine's videprinter. By default
it is switched off. You can switch it on in Display Setup under the
Standings/Videprinter screen options. The default threshold for what
counts as an "unbelievable" bracketable score is 90, which seems a
reasonable value for 9-round Countdown games. This threshold can be
changed in the same place on the Display Setup page.

Note that this only applies to the videprinter display on the public
display screen, and not the "recent games" box on the score entry page.