Automatically upload tourney state to be visible at greem.co.uk/co/<tourney>
Atropine can now be set up to upload the tourney state to the web
automatically every 10 seconds, so that absent spectators can follow
the event.

This requires:
 * A username and password for the tourney-state-submission CGI script
   https://greem.co.uk/cgi-bin/colive/submit.py
 * An internet connection at the event.

The main Python script, atropine.py, now starts an "uploader thread"
whose job is to wake up every second and upload all the tourneys that
need uploading to the web. Generally a tourney will be uploaded if
the operator has requested it on that tourney's Broadcast Setup page,
and if the state for that tourney hasn't been uploaded in the last
10 seconds.

The uploader thread posts a JSON object to the submit.py script on
the server. This JSON object contains standings, games, videprinter
log entries, and so on - everything the webpage needs to know about
the tourney in order to display the tourney's state to viewers.

Once the main program is killed and restarted, its uploader thread
starts up with an empty list of tourneys. So in the event that
atropine.py is restarted while broadcasting is enabled, the user will
have to switch the feature on in Broadcast Setup again.

Getting started

Clicking the "Start Broadcasting" button sends a message to the
uploader thread to tell it to add this tourney to the list of
uploadable tourneys. It will continually upload the state, as JSON,
to the website until either the main program (atropine.py) exits, or
the user tells it to stop ("Stop Broadcasting").

Where it ends up

Live tourney updates for the tourney foobar can be viewed at
https://greem.co.uk/co/foobar - this page updates automatically
using Javascript unless it's been more than 24 hours since the last
update was received.

The traffic light in the sidebar

The upload feature, unlike every other feature in Atropine, obviously
requires internet access. In the sidebar under the "Broadcast Setup"
link there is a traffic light which either reassures you that updates
are still being posted, or alerts you to any problems.

 * If no lights are lit, the feature is not enabled and no attempt is
   made to send updates (this is the default).
 * If the green light is lit, then the feature is enabled, and it
   successfully uploaded the tourney state some time in the last 15
   seconds.
 * If the amber light is flashing, then either it's been more than 15
   seconds since the last upload but we haven't actually got an error
   message yet, or we have not yet successfully uploaded anything since
   the user pressed "Start Broadcasting".
 * If the red and amber lights are flashing, then the last upload
   attempt failed because either we were unable to connect to the
   server or we got some sort of HTTP error. This happens when your
   internet connection flakes out or some problem occurs on the server.
 * If the red light is flashing, then the last upload attempt failed
   because although we successfully submitted the request, we got an
   error back. This error is probably the user's fault (e.g. wrong
   password). Error message details can be viewed on the Broadcast
   Setup page.

Tourney's "unique ID"

The tourney db file now contains a randomly-generated "unique ID",
generated when the tourney is created. This is stored as the
"uniqueid" attribute in the database's OPTIONS table, and included
with the tourney state uploaded to the web. The purpose of this is
to avoid accidentally overwriting last year's CoTrodmore webpage
with this year's, because you happened to call both tourneys
"cotrodmore".

When a tourney is uploaded, if no tourney by that name exists on the
website then it is created. If a tourney already exists by that name,
then if it's the same db file (i.e. with the same unique ID), the
uploaded state replaces what's on the website. If a tourney already
exists by that name with a different unique ID, the upload is
rejected. The operator must either choose a different tourney name,
or explicitly delete the existing one from the web (there's a
button to do this in Broadcast Setup).

A user may only delete, or post updates for, a tourney created by
them. The name of the user who created each tourney is recorded on
the server.

Extra tourney information

The existing "tourney name" may only contain letters, numbers and
underscores (e.g. "colin_xvi"). This name is used to identify the
tourney on the website and to name the tourney's .db file.

The General Setup page now shows, along with the tourney rules displayed
after the player list is created, two text fields for the tourney's full
name (e.g. "Countdown in Lincoln XVI") and its venue (e.g. "The Ark,
North Hykeham"). These will be displayed in preference to the tourney's
(non-full) name at the top of exported tournament reports and the top
of the tourney's webpage. They are stored as the "fullname" and "venue"
attributes in the OPTIONS table.

Videprinter comments

The results entry page's videprinter now contains an edit box and a
button. This is so the operator can enter arbitrary messages to appear
as text on the videprinter, and as news updates on the webpage (if the
tourney is being uploaded). These comments are written to the GAME_LOG
table with a LOG_TYPE of 101, with the text in the new column
COMMENT. Existing LOG_TYPE values are 1 for a new result, or 2 for a
correction to an existing result.

Enormous commit message

This change also comes with an Enormous Commit Message, for everyone's
comfort and safety.