Event date, and option to mark uploaded event as private
The General Setup page for a tourney now contains a field for the event
date. This is sent to the server along with the full name and the venue,
if the broadcast feature is enabled. The date is also included in the
exported tourney reports, and used as the default value for the event
date if you generate Wikitext.

On the Broadcast Setup page you can tick a box to mark an uploaded
tourney as "private". When you visit the (as yet nonexistent) page
on the server to list all tourneys, private tourneys won't appear in
the list. They're only visible to people who happen to have the link.

The tourney rules section of the setup page is now called Tourney
Properties, and has been reorganised. The "draws" checkbox is now
part of the qualification section. The option to rank the standings
table by something other than wins-and-points has moved to the Advanced
page because nobody ever changes it.