Each comment can now be videprinter-only, web-only, both or neither
LOG_TYPE is now 96 to represent a comment, bitwise-ored with 1 if it
should appear on the display's videprinter, and 4 if it should appear
on the web. This means the "old" log_type of 101 means both videprinter
and web.

There are now tick boxes on the comment-posting form saying "to display"
and "to web". These only appear if broadcasting is enabled, otherwise
comments by default are marked as going to both. (If you then enable
broadcasting, all such comments will appear on both.)