Prettify fixture generator menu
The fixture generator menu (the page you see when you click "Generate
Fixtures...") now has a picture associated with each fixture generator,
and a more detailed description for some of the generators.

The descriptions now include nudges for new users, such as "this is a
popular choice for the first round of an event".