Overhaul manual fixgen interface, make finals affect standings
MANUAL FIXTURE GENERATOR INTERFACE NOW LESS AWFUL

The Manual fixture generator interface used to be a grid of drop-down
boxes, each of which contained a list of player names from which to
select. Now it's a grid of text boxes in which the user types in a
player name, and the player name is autocompleted like on the result
entry page. This is a lot easier for data entry.

You can switch back to the drop-down box interface if you want.

The Fully Manual fixture generator used to be a copy-and-paste of the
Manual fixture generator with some bits changed. This is now properly
refactored and the Fully Manual generator calls functions from the
Manual one having first set some attributes in "settings" to indicate
how many groups we are to generate, how many players in each group
(always 2 with Fully Manual) and what type of games to generate.

When the interface for the Fully Manual fixture generator is shown,
you also see the standings table for the relevant division. This is
useful if, as in pretty much all cases when using this fixture
generator, you want to know who came first and second (or possibly
first to fourth, or first to eighth).

Talking of which...

STANDINGS TABLE TAKES ACCOUNT OF FINALS

Furthermore, the standings table now sorts players who play in a
final in order of their performance in the quarter-final (if played),
semi-final (if played) and final.

How it works is that a win, draw or loss in a quarter-final, semi-final
or final is worth a number of "finals points" as follows:

               QF(s=2)    SF(s=1)   Final(s=0)
  Win   (r=3)     48         12         3
  Draw* (r=2)     32          8         2
  Loss  (r=1)     16          4         1

(* Not sure what a draw is even supposed to mean in a finals match, but
there it is anyway.)

In other words, if the result value (r) is 3 for a win, 2 for a draw and
1 for a loss, and the stage value (s) is 0 for a final, 1 for a
semi-final and 2 for a quarter final, the number of "finals points" for
a result in the finals is given by r*4^s. Finals points take precedence
over wins and points when sorting players in the standings table.

This means in a CoJeff three-stage knockout where winners play winners
and losers play losers, WWW > WWL > WLW > WLL > LWW > LWL > LLW > LLL.

Finals points are only awarded to players who played in a final. This
is so that if you play ordinary QF, SF and final rounds where only
winners play on, the winner of the final will have 48+12+3, the loser
of the final will have 48+12+1, and everyone else will have 0.

This means the standings table will only re-sort itself after the
final(s) is/are played. Rejigging the standings table after, say, the
quarter-finals are played just means the standings table won't be very
useful for whoever has to enter the fixtures for the semi-finals and
needs to know who the "1st or 8th" and "4th or 5th" seeds were.

If a player didn't play a quarter-final or a semi-final, then provided
they played a final they're considered to have won the stages they
didn't play. So if some lucky git gets a bye to the final and wins it,
they'll still be ranked above someone who won their QF and SF but lost
the final.

Finals form will appear on the standings table when any QF, SF or final
match is played. On the "standings" page of the operator's interface
it's its own column, and on the public display window it's awkwardly
added on to the "played" column.

There are probably a myriad of ways to break this or make it do odd
things, the most obvious of which is to generate multiple final rounds
in the same division and confuse the standings table by having people
play more than one final. I'm pleading garbage-in-garbage-out on that.

Future work: shoehorn third-place playoffs (game type 3P) into all this.