Make re-sorting standings table by finals outcome optional
Commit dac801032f54355276d02d988e6b217e65cad87b introduced a feature
which (simplified) makes the standings table now puts the winner of
the final at the top, regardless of the number of wins the final
participants previously had.

This commit makes that feature optional. By default it's enabled, and
you can disable it by clearing the new checkbox under "Ranking order"
on the advanced setup page.