Changes to tournament export feature
To save space in the sidebar, there is now only one "Export results"
link, rather than a link for each of the four formats. That one link
takes you to a page which offers you a list of formats.

Also, you can now export the standings as CSV, which you couldn't
before.