Make sidebar less wordy
The links to "Name to Table Index", "Tuff Luck", "Overachievers" and
"Tim Down Award" are now hidden behind a "More" heading, and you have to
click a link to expand the extra links. This saves vertical space in the
sidebar.

Also, the Team Setup table now looks a bit less awful.