Show number of viewers when broadcasting
When the uploader thread submits the game state to the server, it now
gets back a count of the number of distinct visitors to that tourney
over the last minute.

If the last upload attempt was successful, this count, if present, is now
displayed in the sidebar and on the Broadcast Setup page.

Side note: on the server, each time something requests the JSON tourney
state, the server notes in its database the last time that visitor
requested that tourney's state. A "visitor" is identified by their IP
address and HTTP user-agent header, fed through a one-way hash. The
current "viewer count" therefore is the number of distinct visitor
hashes in the database table for that tourney, where the timestamp
of the visit is in the last 60 seconds.