Only upload game state if it's changed since the last upload
This is a bandwidth-saving optimisation. In a large tourney with lots of
games, the JSON object describing the tourney state can reach 100KB or
more. The uploader thread sends this to the server every 10 seconds
whether the tourney state has changed or not. This waste of bandwidth
is particularly problematic if the organiser is tethering their laptop
to their mobile phone's internet connection, for example.

So, for each tourney we're uploading, the uploader thread (uploader.py)
now remembers the last game state object successfully uploaded to
the server, and only sends the server the new state if it's changed
since then.

If it hasn't changed, then there's no point in uploading it all over
again. We still send a message to the server (to submit.py) but it
contains a null or nonexistent "state" attribute, and the rest of the
message is tiny - it contains not much more than the tourney name,
username and password.

When the server receives a submission with a null "state" attribute, it
goes through all the steps it would normally go through when recording
an update to a tourney (authenticate user, check tourney name is valid,
check user owns tourney, include the current number of viewers in the
response), except that it doesn't update its idea of the tourney state,
it just leaves the old file as it is.

Additionally, if the state attribute is null, the server checks that the
tourney exists. If it doesn't, that's an error.