fixturegen.py initialises fixgen_settings in the wrong place
Say you already have Round Robin fixtures generated for Division A for
rounds 1-7, and after generating rounds 1-3 for Division B with (say)
the Random fixture generator, you decide to generate fixtures in
Division B for round 4 using the Swiss generator.

You'll find you can generate fixtures, but when you press Accept, you
get an error message complaining that Division A (which you aren't
generating fixtures for!) has unplayed games, so you can't use the
Swiss fixture generator.

This will pose a problem for the FOCAL Finals unless fixed - apparently
generating fixtures for only a subset of divisions is a feature that
only gets tested once per year.

This was caused by commit 6258fb114e9f8dc18b675b05948a6dd8f7255a68.
After fixturegen.py acts on whatever it's been told to do, but before it
displays anything to the user, we incorrectly reinitialise
fixgen_settings, which contains, among other things, the details of
which divisions we're generating fixtures for. This means when we call
show_fixtures_to_accept(), the HTML form we present to the user hasn't
got this information.

So when the user presses Accept, this sends a request containing only
the list of fixtures to accept but no information about which divisions
are having fixtures generated in them. The bit of code which takes the
fields in the submitted form, puts them in fixgen_settings, and then
calls the fixgen's check_ready() method therefore assumes we're
generating fixtures for ALL divisions (as it would if we had only just
selected which fixgen we wanted to use, for example), and sets
fixgen_settings accordingly. The Swiss fixgen's check_ready() method
therefore fails, because it thinks it's generating fixtures for Division
A and B, and Division A has unplayed games in it.

Solution: move the initialisation of fixgen_settings to the start of
the script where it should be. Before check_ready() is called, we
populate fixgen_settings with the fields from the submitted form (if
any), and then when we output the HTML page to the user, if we need to
refer to fixgen_settings we use that populated object.