Make uploader interval period a global variable in uploader.py
This variable is uploader_interval_sec, and defaults to 10. The uploader
thread uploads the current tourney state to the server (if it's changed
since the last upload) every this many seconds.

Moved to a global variable at the top of the file so that it's more
easily changeable if anyone wants to change it.