Remove all uses of cgi.escape()
The cgi module's escape() function has been deprecated for a while and
is gone from in Python 3.8 onwards. This causes AttributeError
exceptions to be thrown if you run Atropine under Python 3.8.

Replace all remaining calls to cgi.escape with calls to cgicommon.py's
escape() function, which calls the recommended replacement function
html.escape().

Thanks to Callum Todd for finding this.