Rename Fully Manual fixgen to Raw, and add accessibility note
I kept clicking the "Fully Manual" fixture generator when I wanted
"Manual Pairings/Groups". To prevent this happening in future, I've
renamed the "Fully Manual" fixture generator to "Raw".

This means someone in a hurry looking down the menu for the one that
says "manual" will always find the "Manual Pairings/Groups" generator,
which is usually the one people want.

Also, for the Manual Pairings/Groups generator, if there are any
accessible tables, the interface now adds a note at the bottom warning
the user that the displayed table numbers may be reassigned to fulfil
accessibility requirements. A wheelchair symbol is displayed next to
any table numbers designated as "accessible".