Add "final between top two" fixture generator
By far the most common reason to use the Raw fixture generator is to
generate a final between the top two players in the standings table.
So let's have a dedicated fixture generator that automates this
rather than making the user pick the top two players from a list.

The "Final between top two" fixture generator will create a new round,
and for each division, a final (type F) between the top two players in
that division's standings table.

This fixture generator only works if there are exactly two players who
are unambiguously the "top two". If there are three or more players tied
for first place, or two or more players tied for second place, you'll
get an error message explaining that you'll need to rule on who goes
through to the final and use the Raw fixture generator to specify the
players manually.