For the Hangover team score display, use black text on the white box
Currently we use white text on the Hangover team score boxes, including
the one representing the white team, which has a white background. The
text has a black shadow around it, so it's still visible on the white
background, but it would be better to simply use black text if the
background is white.

So if the team colour is a "light" colour (using a rather unscientific
way of deciding this) then the team score box should have black text
with no shadow, and if it's a "dark" colour then it should have white
text with a black shadow as before.

The default team colours are blue and white, as before. The particular
shade of blue we use is considered a "dark" colour, and the white is
obviously a "light" colour.