Fixture generator menu: context-sensitive suggestions at the top
Currently, the fixture generator menu has eight fixture generators in
it, and they are sorted alphabetically by name. This means, among other
things, that the Swiss Army Blunderbuss fixture generator, which is by
far the most popular, is at the bottom of the list, and even on a
reasonably-sized screen the user has to scroll down to find it.

So let's change the fixture generator menu so it has a few selected
"suggested" fixture generators at the top, and then the others in a
separate list below.

Which fixture generators are "suggested" depends on when in the tourney
we are.

If we haven't generated any rounds yet, the suggested generators should
be those often used to generate round 1: both Random generators, Round
Robin, and Manual.

If we have already generated round 1, we should suggest the generators
most often used to generate the second round onwards: Swiss, Random and
Final. There's not really much to gain from alphabetically sorting a
list of three, so present this suggested list in that order: Swiss first
as the most commonly-used, then Random, then Final (the least-used, and
also conveniently the one which if used will get used last in the
tourney).

In either case, nothing is "hidden" or "gone"; the non-suggested fixture
generators are shown in alphabetical order below the suggested list,
just as before.