Fixture generation: no need to show division name if there's only one
When you select a fixture generator and it asks you how many players are
going to go on each table for each division, it puts e.g. "Division A
(24 players)" in the heading.

But if there is only one division, then from the user's point of view
divisions aren't even a thing, so let's not confuse them by wittering
about "Division A".