Improve fixgen error handling with unusual player counts
For the raw fixture generator, it's valid to put the same player in
multiple slots, so we shouldn't flag that as an error.

Also, if the tourney has an awkward number of players (e.g. 7) then
most of the fixture generators (other than Raw or Round Robin) should
immediately return an error from their check_ready() methods saying
that the number of players in the division is not compatible with any
supported table size. The Manual fixture generator instead did something
weird - presenting the user an empty list of table sizes to pick from.

If there is only one player in the tournament, the only fixture
generator that can do anything even remotely useful is the Raw
generator, which in those circumstances can only generate games where
the player plays against themselves (this is allowed!).