Auto mode to show Standings/Results view when final is generated
The method used by the standings/results view to choose which round's
games are shown has been subtly changed.

The present behaviour is to show the latest round to contain at least
one completed game. If there is no such round it shows round 1, and if
there is no round 1 it shows nothing.

The new behaviour is that if the round after the last completed round
contains exactly one game, it shows that (one-game) round. Otherwise,
the behaviour is as before.

This is so that when you generate a final, the automatic display now
switches to the standings/results view, showing the standings and the
final. This is preferable to the previous behaviour of showing the
fixtures view like when ordinary rounds are generated, which would take
up the whole screen showing one game at a stage in the tournament when
everyone wants to know where they finished in the standings.