Add help link in sidebar which leads to documentation
Put the documentation in docs/ into webroot/docs, which is really where
it should have been all along. This means Atropine's web server can serve
it, and we can have a link in the sidebar that opens it in a new window,
rather than expecting the user to navigate to the docs folder in their
OS's file manager and open it that way, which I'm betting nobody ever
did.

Bump version number to 1.1.4.