Improve fixture generator players-per-table selection
When the user has asked to generate fixtures, make it more obvious if
the number of active players isn't a multiple of 2 or 3. Put up a
warning if this is the case.

Also put the number of active players in the text of the page which
shows the "how many players per table" control, not in the heading where
people ignore it.