Edit fixtures page can now add or delete individual fixtures
The "edit fixtures" page, reachable from the main score entry page for a
round, now has a checkbox for each division which, if checked, reveals
an additional set of form elements to add a new game to the round. This
allows additional matches to be added to the round one by one if this
is required after the fixtures have been generated and announced.

Individual fixtures may now also be deleted by checking the delete box
next to the relevant fixture before saving changes. The "delete all
games in this fixture" checkbox is gone, and replaced with a pair of
buttons to check all or uncheck all the delete boxes in a division.