Remove livegame/livegame.py
livegame.py is the script that was used many years ago to show a
round-by-round display of a live game, usually a final. It used pygame
to display the graphics, which Atropine doesn't depend on any more, and
livegame.py was never even ported to Python 3. It hasn't been used for
years and isn't likely to be, so it's being removed.