Experimental solution to the Mac #! line problem
On a Mac, Python 3 is not usually at /usr/bin/python3, but in some
custom location. This means all the #! lines at the start of the Python
scripts, which are required for the CGI scripts run by the web server,
are usually wrong on a Mac.

Even worse, if you try putting "#!/usr/bin/env python3" at the top of
every file, which works on Linux, the Mac treats it as all one argument
and complains that it can't find an interpreter called "env python3"
in /usr/bin.

However, we do have the advantage that modern Macs are basically Unix
systems in disguise, so you've got access to bash and related utilities.

So, macify.sh is a bash script which works out where your Python 3
interpreter is ("which python3") and uses sed to replace the #! line
at the top of every *.py file in an Atropine installation with a line
pointing to the correct location of the Python 3 interpreter.

Now all we need is a Mac to test it on.